Moves ctl kubeconfig flags to top level
