+++ .gitignore
@@ -3,6 +3,7 @@
 /acmesolver
 /controller
 /ingress-shim
+/ctl
 /hack/build/dockerfiles/cert-manager-*_*_*
 .vscode
 .venv
