+++ cmd/ctl/BUILD.bazel
@@ -5,16 +5,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "github.com/jetstack/cert-manager/cmd/ctl",
     visibility = ["//visibility:private"],
-    deps = [
-        "//cmd/ctl/cmd:go_default_library",
-        "//pkg/util/cmd:go_default_library",
-<<<<<<< HEAD
-        "@io_k8s_klog//:go_default_library",
-=======
-        "@com_github_spf13_cobra//:go_default_library",
-        "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
->>>>>>> 65e8ff5db... Cleans up renew kube flags and adds --namespace validation
-    ],
+    deps = ["//pkg/util/cmd:go_default_library"],
 )
 
 go_binary(
