+++ cmd/ctl/cmd/BUILD.bazel
@@ -12,6 +12,7 @@ go_library(
         "@com_github_spf13_cobra//:go_default_library",
         "@io_k8s_cli_runtime//pkg/genericclioptions:go_default_library",
         "@io_k8s_client_go//plugin/pkg/client/auth:go_default_library",
+        "@io_k8s_klog//:go_default_library",
         "@io_k8s_kubectl//pkg/cmd/util:go_default_library",
     ],
 )
